+++ packages/ui/core-components/src/lib/unsorted/viz/map/Map.stories.svelte
@@ -44,7 +44,7 @@
 	);
 </script>
 
-<!-- Exlcuded from chromatic, map layers don't reliably load in the same order -->
+<!-- Excluded from chromatic, map layers don't reliably load in the same order -->
 <Story name="Basic Usage" parameters={{ chromatic: { disableSnapshot: true } }}>
 	<BaseMap title="My Map" height="300">
 		<Points