+++ packages/ui/icons/index.js
@@ -3,7 +3,7 @@ import { ThemeBuilder } from '@steeze-ui/icons';
 const builder = new ThemeBuilder({
 	sources: {
 		// inputRaw: './custom-icons/', //* use this if your sources come from a npm package
-		// collectFromDir: { 'custom-icons': 'default' }, //* use this if icon themes are bound to the overlying derectory (e.g. outline/cake.svg, fill/cake.svg).
+		// collectFromDir: { 'custom-icons': 'default' }, //* use this if icon themes are bound to the overlying directory (e.g. outline/cake.svg, fill/cake.svg).
 		// collectFromSuffix: { '-line': 'default', fill: 'solid' }, //* use this if icon themes are bound to a suffix (e.g. cake-line.svg, cake-fill.svg).
 		// fallbackTheme: 'default', //* where to collect, when icons cannot be identified by their suffix (default is dont collect)
 	},
@@ -13,4 +13,4 @@ const builder = new ThemeBuilder({
 	}
 });
 
-builder.build(); //* if neither collectFromDir or collectFromSuffix are defined, a already populated themes directory with at least a default theme is necessary (see example in this themplate)
+builder.build(); //* if neither collectFromDir or collectFromSuffix are defined, a already populated themes directory with at least a default theme is necessary (see example in this template)